@@ -55,13 +80,13 @@
     "prosemirror-state": "^1.4.2",
     "prosemirror-transform": "^1.7.1",
     "prosemirror-view": "^1.30.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-helmet-async": "^1.3.0",
-    "react-math-view": "^1.3.2",
-    "react-scripts": "5.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-map-gl": "^7.1.3",
     "sass": "^1.59.2",
-    "typescript": "4.0.3",
+    "tippy.js": "^6.3.7",
+    "tiptap-markdown": "^0.8.2",
+    "typescript": "5.1.6",
     "typia": "^3.6.8",
     "uuid": "^9.0.0",
     "web-vitals": "^3.1.0",
