@@ -1,3 +1,5 @@
+'use client'
+
 import React from "react";
 import { IndexeddbPersistence } from "y-indexeddb";
 import { TiptapCollabProvider } from '@hocuspocus/provider'
@@ -7,23 +9,28 @@ import { QiC, QiId, QiT } from "../core/Model";
 export const QiStoreContext = React.createContext<QiT>(new QiC());
 
 export const QiStore = (props: { qiId: QiId, userId: string, children: JSX.Element}) => {
-  // Initialise an empty ydoc to fill later with data from IndexedDB
+  // Initialise an empty yDoc to fill with data from TipTap Collab (online) and IndexedDB (offline)
   const qi = new QiC()
 
+  // Anyone accessing this particular "room" will be able to make changes to the doc
+  // The room can also be understood to be the unique id of each qi
   const roomName = props.qiId
 
   const appId = 'dy9wzo9x'
 
-  // Persist the document using the cloud provider
+  //  Sync the document locally
+  new IndexeddbPersistence(roomName, qi.information)
+
+  // Sync the document using the cloud provider
   new TiptapCollabProvider({ 
     appId: appId,// get this at collab.tiptap.dev
     name: roomName, // e.g. a uuid uuidv4();
-    token: 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpYXQiOjE2ODQxNDQ5MDAsIm5iZiI6MTY4NDE0NDkwMCwiZXhwIjoxNjg0MjMxMzAwLCJpc3MiOiJodHRwczovL2NvbGxhYi50aXB0YXAuZGV2IiwiYXVkIjoia29uZ3dlaUBldXNheWJpYS5jb20ifQ.bUsMJ8W_T15zk0PWdiBddMeVLNyppDlI6g7Vr3dIA3s', // see "Authentication" below
+    token: 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpYXQiOjE2OTE0NTk2MDIsIm5iZiI6MTY5MTQ1OTYwMiwiZXhwIjoxNjkxNTQ2MDAyLCJpc3MiOiJodHRwczovL2NvbGxhYi50aXB0YXAuZGV2IiwiYXVkIjoia29uZ3dlaUBldXNheWJpYS5jb20ifQ.AeaTq-IkPZ5n21XAVj5lS5IfUNNsz3As25sUhkexxQY',
     document: qi.information
   });
 
-  //  Persist the document locally
-  new IndexeddbPersistence(roomName, qi.information)
+  console.log("roomName", roomName)
+
 
   return (
     <QiStoreContext.Provider value={qi}>
