@@ -1,36 +1,34 @@
 import React from "react";
 import { Node, NodeViewProps, wrappingInputRule } from "@tiptap/core";
 import { NodeViewContent, NodeViewWrapper, ReactNodeViewRenderer, nodeInputRule } from "@tiptap/react";
-import { Group } from "./Group";
-import { Qi } from "../../core/Qi";
-import { group } from "console";
+import { motion } from "framer-motion";
+import { parchment, purple } from "../Theme";
 
-// TODO: Match for brackets with text in between
-export const groupInputRegex = /{([^{}]*)}/;
+const REGEX_BLOCK_SLASH = /\/\/[^/]+\/\//
 
-export const GroupExtension = Node.create({
-  name: "group",
+export const CommentExtension = Node.create({
+  name: "comment",
   group: "block",
   content: "block*",
   // TODO: Doesn't handle inline groups
   inline: false,
-  selectable: true,
-  draggable: true,
+  selectable: false,
   atom: true,
   parseHTML() {
     return [
       {
-        tag: "group",
+        tag: "comment",
       },
     ];
   },
   renderHTML({ HTMLAttributes }) {
-    return ["group", HTMLAttributes, 0];
+    return ["comment", HTMLAttributes, 0];
   },
+  draggable: true,
   addInputRules() {
     return [
       wrappingInputRule({
-        find: groupInputRegex,
+        find: REGEX_BLOCK_SLASH,
         type: this.type,
       })
     ]
@@ -39,9 +37,13 @@ export const GroupExtension = Node.create({
     return ReactNodeViewRenderer((props: NodeViewProps) => {
       return (
         <NodeViewWrapper>
-          <Group lens={"verticalArray"} qid={props.node.attrs.qid}>
-            <NodeViewContent />
-          </Group>
+          <motion.div style={{
+            backgroundColor: parchment, borderRadius: 5, padding: `20px 20px 20px 20px`, color: "#343434"
+          }}>
+            <div style={{fontFamily: "EB Garamond", fontSize: 18}}>
+              <NodeViewContent />
+            </div>
+          </motion.div>
         </NodeViewWrapper>
       );
     });
